requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==1.4.46
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
